PR review cleanup
